Add files via upload
As discussed with Leads , I have written a utility which is used in separating the file names according to extension. It helps doc team while adding the file names into installation document.

In upcoming release we may ask to do it manually, so to save time and reduce the error, we have the small utility in java.